Modularize scripts
